Upload journal updater lambda function
A lambda function that reads a searches a Google Document for a journal from today's date and uploads it to DynamoDB. Meant to be used with an Event Scheduler. Does not work without a service-a file for google cloud service account given view access to Google Doc. Document ID left blank.